Chore: 파일 사이즈 단위를 변환해주는 함수를 모듈화
기존 list.js 에서 파일 사이즈 단위를 변환해주기 위해서 사용하던
bytesToSize() 함수를 zlib.js에서도 재사용하기 위해서
따로 bytesToSize.js 파일로 분리해서 모듈화하였다.
