+++ bytesToSize.js
@@ -1,0 +1,10 @@
+// 파일 크기 단위를 바꿔주는 함수
+function bytesToSize(bytes) {
+	const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB'];
+	if (bytes === 0) return 'n/a';
+	const i = parseInt(Math.floor(Math.log(bytes) / Math.log(1024)), 10);
+	if (i === 0) return `${bytes} ${sizes[i]}`;
+	return `${(bytes / 1024 ** i).toFixed(1)} ${sizes[i]}`;
+}
+
+module.exports = { bytesToSize };
